backend/requirements.txt
fastapi
openai
pydantic
python-dotenv
python-multipart
supabase
uvicorn